Fix gold_commuter_hubs to count all network nodes
The `gold_commuter_hubs` transformation calculated an `indegree` result
set of all station_ids where a given id was the ending station, then
calculated an `outdegree` set of all station_ids where a given id was the
starting station. The final step joined the two result sets to combine
and calculate the total indegree and outdegree of each station.

However, by using a default `JOIN`, which is an inner join, we only
included nodes that were both a starting node _and_ an ending node.
This choice ommited all true source and sink nodes from the network,
where a station_id was found in only one of the two sets.

What is required here is a full outer join, which in turn requires some
care because the result for sink and source nodes will have only one
non-null station_id and only one set of non-null values for in/out
degree and weighted in/out degree.